feat: add about API for fetching members and update localization files
- Implemented `fetchMembersServer` function to retrieve and normalize member data from Directus. jsut moved the funciton to the new entity/about folder
- Created API index for easier exports in the About entity.
- Added types for Member interface.
- Updated all json files with the new count value
