+++ frontend-next-migration/src/app/[lng]/(helper)/about/_getPage.ts
@@ -1,7 +1,7 @@
 import { useServerTranslation } from '@/shared/i18n';
 import { createPage } from '@/app/_helpers';
 import { AboutPageProps } from '@/preparedPages/AboutPage';
-import { fetchMembersServer } from '@/entities/Member';
+import { fetchMembersServer } from '@/entities/About';
 
 export async function _getPage(lng: any) {
     const { t } = await useServerTranslation(lng, 'about');
@@ -25,9 +25,9 @@ export async function _getPage(lng: any) {
             nationality: t('nationality'),
             behind: t('behind'),
             projectCount: uniqueMemberCount,
-            localityCount: '15',
-            nationalityCount: '5',
-            behindCount: '6',
+            localityCount: t('localityCount'),
+            nationalityCount: t('nationalityCount'),
+            behindCount: t('behindCount'),
             V2019: t('V2019'),
             V2020: t('V2020'),
             V2021: t('V2021'),
