+++ frontend-next-migration/src/entities/About/api/aboutApi.ts
@@ -1,0 +1,29 @@
+import { envHelper } from '@/shared/const/envHelper';
+import { createDirectus, rest, readItems } from '@directus/sdk';
+import { Member } from '../model/types/types';
+
+const directusBaseUrl = envHelper.directusHost;
+const client = createDirectus(directusBaseUrl).with(rest());
+
+export async function fetchMembersServer(): Promise<number> {
+    try {
+        const members = await client.request<Member[]>(readItems('members', { limit: 500 }));
+
+        const normalize = (s: unknown) =>
+            String(s ?? '')
+                .normalize('NFKD')
+                .replace(/[\u0300-\u036f]/g, '')
+                .replace(/\s+/g, ' ')
+                .trim()
+                .toLowerCase();
+
+        return new Set(
+            members.map((m, i) =>
+                normalize(m.name) ? `name:${normalize(m.name)}` : `__id_${m.id ?? i}`,
+            ),
+        ).size;
+    } catch (error) {
+        console.error('fetchMembersServer error:', error);
+        throw error;
+    }
+}
