+++ frontend-next-migration/src/entities/About/api/index.ts
@@ -1,0 +1,1 @@
+export { fetchMembersServer } from './aboutApi';
