+++ frontend-next-migration/src/entities/About/index.ts
@@ -1,0 +1,1 @@
+export { fetchMembersServer } from './api';
