+++ frontend-next-migration/src/entities/About/model/types/types.ts
@@ -1,0 +1,4 @@
+export interface Member {
+    id: number;
+    name: string;
+}
