+++ frontend-next-migration/src/entities/Member/api/index.ts
@@ -1,1 +1,1 @@
-export { useGetMembersQuery, fetchMembersServer } from './membersApi';
+export { useGetMembersQuery } from './membersApi';
