+++ frontend-next-migration/src/entities/Member/ui/MemberItem.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import Image from 'next/image';
 import Link from 'next/link';
